Refactor: DocumentAPI -> documentAPI 이름 변경
